+++ src/services/PostmanApiService.ts
@@ -12,14 +12,14 @@ export class PostmanApiService {
     this.apiKey = `${process.env.POSTMAN_API_KEY}`
   }
 
-  async createCollection(collection: CollectionDefinition): Promise<string> {
+  async createCollection(collection: CollectionDefinition, workspaceId?: string): Promise<string> {
     const data = JSON.stringify({
       collection: collection
     })
-
+    const workspaceIdParam = workspaceId ? `?workspace=${workspaceId}` : ''
     const config = {
       method: 'post',
-      url: `${this.baseUrl}/collections`,
+      url: `${this.baseUrl}/collections${workspaceIdParam}`,
       headers: {
         'Content-Type': 'application/json',
         'X-API-Key': this.apiKey
@@ -52,14 +52,18 @@ export class PostmanApiService {
     }
   }
 
-  async updateCollection(collection: CollectionDefinition, uuid: string): Promise<string> {
+  async updateCollection(
+    collection: CollectionDefinition,
+    uuid: string,
+    workspaceId?: string
+  ): Promise<string> {
     const data = JSON.stringify({
       collection: collection
     })
-
+    const workspaceIdParam = workspaceId ? `?workspace=${workspaceId}` : ''
     const config = {
       method: 'put',
-      url: `${this.baseUrl}/collections/${uuid}`,
+      url: `${this.baseUrl}/collections/${uuid}${workspaceIdParam}`,
       headers: {
         'Content-Type': 'application/json',
         'X-API-Key': this.apiKey
@@ -129,9 +133,111 @@ export class PostmanApiService {
             return <any>new Date(b.updatedAt) - <any>new Date(a.updatedAt)
           })
           console.log(
-            '\nMultiple Postman collection matching "' +
-              collName +
-              '", the most recent collection is updated.'
+            `\nMultiple Postman collection matching "${collName}", the most recent collection is updated.`
+          )
+          match = matches[0]
+        }
+      }
+      return match
+    } catch (error) {
+      console.log(error?.response?.data)
+      return error.toString()
+    }
+  }
+
+  async findWorkspaceCollectionByName(
+    workspaceId: string,
+    collName: string
+  ): Promise<CollectionDefinition> {
+    const config = {
+      method: 'get',
+      url: `${this.baseUrl}/workspaces/${workspaceId}`,
+      headers: {
+        'Content-Type': 'application/json',
+        'X-API-Key': this.apiKey
+      }
+    } as AxiosRequestConfig
+
+    try {
+      const res = await axios(config)
+      const data = res.data
+      let match = {}
+
+      if (data?.workspace.collections) {
+        // Match all items by name, since Postman API does not support filtering by name
+        const matches = data.workspace.collections.filter((o: CollectionDefinition) => {
+          if (!o?.name) return
+          return (
+            o.name.toLowerCase().replace(/\s/g, '') === collName.toLowerCase().replace(/\s/g, '')
+          )
+        })
+
+        if (matches.length === 1) {
+          match = matches[0]
+        }
+        if (matches.length > 1) {
+          // Sort by date and take newest
+          matches.sort((a, b) => {
+            // Turn your strings into dates, and then subtract them
+            // to get a value that is either negative, positive, or zero.
+
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            return <any>new Date(b.updatedAt) - <any>new Date(a.updatedAt)
+          })
+          console.log(
+            `\nMultiple Postman collection matching "${collName}" in the workspace, the most recent collection is updated.`
+          )
+          match = matches[0]
+        }
+      }
+      return match
+    } catch (error) {
+      console.log(error?.response?.data)
+      return error.toString()
+    }
+  }
+
+  async findWorkspaceByName(workspaceName: string): Promise<CollectionDefinition> {
+    const config = {
+      method: 'get',
+      url: `${this.baseUrl}/workspaces`,
+      headers: {
+        'Content-Type': 'application/json',
+        'X-API-Key': this.apiKey
+      }
+    } as AxiosRequestConfig
+
+    try {
+      const res = await axios(config)
+      const data = res.data
+      let match = {}
+
+      if (data.workspaces) {
+        // Match all items by name, since Postman API does not support filtering by name
+        const matches = data.workspaces.filter(
+          (o: { id?: string; name?: string; type?: string }) => {
+            if (!o?.name) return
+            return (
+              o.name.toLowerCase().replace(/\s/g, '') ===
+              workspaceName.toLowerCase().replace(/\s/g, '')
+            )
+          }
+        )
+
+        if (matches.length === 1) {
+          match = matches[0]
+        }
+        if (matches.length > 1) {
+          // Sort by date and take newest
+          matches.sort((a, b) => {
+            // Turn your strings into dates, and then subtract them
+            // to get a value that is either negative, positive, or zero.
+
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            return <any>new Date(b.updatedAt) - <any>new Date(a.updatedAt)
+          })
+          console.log(
+            `\nMultiple Postman workspaces matching "${workspaceName}", the most recent workspace is updated.`
           )
           match = matches[0]
         }