+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Added support for Postman Workspace when uploading collections (#118)
 - Bug fix where portmanReplacements & upload to Postman had different results (#130)
 - Improve output for CI and smaller terminals (#126)
 - Updated example